Устранил замечания к заданию 3 из модуля 7
