+++ project/package.json
@@ -6,6 +6,7 @@
     "@reduxjs/toolkit": "^1.5.0",
     "axios": "^0.21.1",
     "classnames": "^2.3.1",
+    "dayjs": "1.11.5",
     "history": "^5.2.0",
     "http-status-codes": "2.2.0",
     "leaflet": "^1.8.0",
