+++ project/src/components/header/header.tsx
@@ -1,13 +1,17 @@
 import { HeaderLogo } from '../header-logo/header-logo';
 import { UserNavigation } from '../user-navigation/user-navigation';
 
-export function Header(): JSX.Element {
+type HeaderProps = {
+  userNavigation?: boolean;
+}
+
+export function Header({userNavigation = true}: HeaderProps): JSX.Element {
   return (
     <header className="header">
       <div className="container">
         <div className="header__wrapper">
           <HeaderLogo/>
-          <UserNavigation/>
+          {userNavigation ? <UserNavigation/> : ''}
         </div>
       </div>
     </header>
