+++ project/src/components/room-gallery/room-gallery.tsx
@@ -6,7 +6,9 @@ type RoomGalleryProp = {
 
 export function RoomGallery({ images, type }: RoomGalleryProp): JSX.Element {
 
-  const renderImages = () => images?.map((image) => (
+  const imageForRender = images?.slice(0, 6);
+
+  const renderImages = () => imageForRender?.map((image) => (
     <div key={`picture-${image}`} className="property__image-wrapper">
       <img className="property__image" src={image} alt={type}/>
     </div>
